feat(commission): implement commission settings retrieval and update routes
