+++ server/models/employeeModel.ts
@@ -645,20 +645,6 @@ const getAllEmployeesForDropdown = async () => {
   }
 };
 
-const getAllCommissionSettings = async () => {
-  try {
-    const query = `
-      SELECT id, key, value FROM settings
-      WHERE type = 'Commission'
-    `;
-    const result = await pool().query(query);
-    return result.rows;
-  } catch (error) {
-    console.error('Error fetching commission settings list:', error);
-    throw new Error('Error fetching commission settings list');
-  }
-};
-
 const getAllRolesForDropdown = async () => {
   try {
     const query = `
@@ -920,7 +906,6 @@ export default {
   updateEmployeePassword,
   getAllEmployees,
   getAllEmployeesForDropdown,
-  getAllCommissionSettings,
   createSuperUser,
   getUserCount,
   getUserData,
